RandomLMs/pyproject.toml
[tool.poetry]
name = "torch"
version = "1.0.0"
description = ""
authors = ["Sounish Nath <flock.sinasini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.11"
pyspark = "^3.4.0"
google-cloud-storage = "^2.9.0"
google-auth = "^2.19.1"
langchain = "^0.0.189"
google-cloud-aiplatform = "^1.25.0"
fuzzywuzzy = "^0.18.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
